Add files via upload
It highlights the vision of the company, and I focused on using Pandas, Seaborn, and MATLAB to present the data graphically. Furthermore, NumPy was essential for numerical calculations. Overall, this is not the actual file; since I left the job, I must keep it hidden for security purposes. Additionally, I used Power BI and SQL to analyze the existing database.